LinearAlgebra divided
