+++ app/src/main/java/com/example/geometry/GUI/Distance.java
@@ -1,0 +1,15 @@
+package com.example.geometry.GUI;
+
+import com.example.geometry.FigureModel.Node;
+
+public class Distance{
+    public float dist;
+    public Node node;
+
+    public Distance() { }
+
+    public Distance(float dist, Node node) {
+        this.dist = dist;
+        this.node = node;
+    }
+}
